+++ bakedanuki/bakedanuki-util/python/bd_util/maya/node/operator/node/dg/anim_layer.py
@@ -1,8 +1,9 @@
 # coding: utf-8
 from ._generated.anim_layer import GeneratedAnimLayer
+from ._anim_layer import AnimLayerOperations
 
 
-class AnimLayer(GeneratedAnimLayer):
+class AnimLayer(GeneratedAnimLayer, AnimLayerOperations):
     __slots__ = ()
 
     NODE_TYPE = "animLayer"